[#28] refactor: ProductionContainer 서비스 빈으로 변경
- ProductContainer.java: component => service 디렉토리로 이동

- 기타: import 패키지 경로 수정
