+++ src/main/java/com/toy/getyourfriday/service/ScrapingManager.java
@@ -1,6 +1,5 @@
 package com.toy.getyourfriday.service;
 
-import com.toy.getyourfriday.component.ProductContainer;
 import com.toy.getyourfriday.domain.scraping.ModelUrl;
 import com.toy.getyourfriday.domain.scraping.WebScraper;
 import org.openqa.selenium.WebDriver;
