+++ src/test/java/com/toy/getyourfriday/component/ProductContainerTest.java
@@ -3,6 +3,7 @@ package com.toy.getyourfriday.component;
 import com.toy.getyourfriday.domain.product.Product;
 import com.toy.getyourfriday.domain.product.Products;
 import com.toy.getyourfriday.domain.scraping.ModelUrl;
+import com.toy.getyourfriday.service.ProductContainer;
 import com.toy.getyourfriday.service.UpdateService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
