+++ src/test/java/com/toy/getyourfriday/domain/scraping/WebScraperTest.java
@@ -1,8 +1,8 @@
 package com.toy.getyourfriday.domain.scraping;
 
 import com.toy.getyourfriday.component.ModelUrlParser;
-import com.toy.getyourfriday.component.ProductContainer;
 import com.toy.getyourfriday.config.ApplicationConfig;
+import com.toy.getyourfriday.service.ProductContainer;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.chrome.ChromeDriver;
